fix(mailroom): keep the server-actions module to async exports — parser and result types move to lib/mailroom/mailbox-form (BI-727D5FD9)
The production build refused actions.ts because it exported a synchronous
parser and types (Next.js: Server Actions must be async functions; AGENTS.md
section 6). parseMailboxForm, formField and the two result types now live in a
plain module imported by the actions and the client components.
